requirements.txt
google-generativeai
python-dotenv
discord.py
